refactors ping job poster to make it sticky
Previously ping jobs where distributed on a round-robin basis, which led
to many prometheus "out-of-order samples" errors.

This new version affects a host to an agent and sticks to it as long as
possible. Upon agent add or removal, the hosts are redistributed so that
each agent have roughly the same number of hosts to ping.