filters indexed measures without index metric in metric list
From time to time, agents reject pollong job with Bad Request error. A
log inspection has showed that it is related to indexed measures that
does not contain IndexMetricID in their measure list.

It also happens to measures whose metrics are polled at periods longer
than the device polling frequency.

This patch is a temporary measure while looking for root cause of this issue.